# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Yasser AI Academy API
  version: 1.0.0
  description: Authenticated education, commerce, portal, tutor, and credential APIs.
servers:
  - url: http://localhost:3001/api
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: yaa_session
...
paths:
...
  /auth/request-email-verification:
    post:
      security: [{ sessionCookie: [] }]
      summary: Send a six-digit, expiring email verification code
      responses:
        "200": { description: Code dispatched or account already verified }
        "401": { description: Authentication required }
        "429": { description: Request limit exceeded }
  /auth/verify-email:
    post:
      security: [{ sessionCookie: [] }]
      summary: Consume a one-time email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp]
              properties:
                otp: { type: string, pattern: '^\d{6}$' }
      responses:
        "200": { description: Email verified }
        "400": { description: Invalid, expired, used, or attempt-limited code }
...
  /tutor/message:
    post:
      security: [{ sessionCookie: [] }]
      summary: Request a guarded Socratic hint
      responses:
        "200": { description: Tutor response }
        "422": { description: Sensitive or moderated content rejected }
...
  /uploads/{purpose}:
    post:
      security: [{ sessionCookie: [] }]
      summary: Upload a signature-validated private asset
      parameters:
        - in: path
          name: purpose
          required: true
          schema: { type: string, enum: [avatar, project_artifact, lesson_media, lesson_download] }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201": { description: Private media asset created }
        "403": { description: Purpose not allowed for this role }
        "413": { description: File exceeds configured maximum }
        "415": { description: Media type or signature rejected }
  /account/uploads:
    get:
      security: [{ sessionCookie: [] }]
      summary: List owned uploads without exposing storage keys
      responses:
        "200": { description: Owned media assets }
  /uploads/{id}:
    get:
      security: [{ sessionCookie: [] }]
      summary: Read an owned private upload
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: File bytes or a temporary provider redirect }
        "404": { description: Asset not found }
    delete:
      security: [{ sessionCookie: [] }]
      summary: Delete an owned upload
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Asset deleted }
...